feat: Relocation de détection d'erreurs dans le modèle et pas dans le contrôleur

+++ app/models/joints/deck_card_remove.js
@@ -4,6 +4,26 @@ var DeckPartsEnum = require.main.require('./app/models/enums/deck_parts');
 
 async function removeCardFromDeck(deckId, cardId, deckPart, userId, callback) {
     logger.debug("Méthode models/joints/deck_card_remove/removeCardFromDeck");
+    
+    err = []
+
+    if (typeof cardId === "undefined" || cardId == null || cardId == "") {
+        err.push("Pas de carte fournie.");
+    }
+    if (typeof userId === "undefined" || userId == null || userId === "") {
+        err.push("Pas d'utilisateur connecté.");
+    }
+    if (typeof deckId === "undefined" || deckId == null || deckId === "") {
+        err.push("Pas de deck fourni.");
+    }
+    if (typeof deckPart === "undefined" || deckPart == null || deckPart === "") {
+        err.push("Pas de partie du deck fournie.");
+    }
+
+    if (err.length > 0) {
+        callback(err.join('<br>'));
+        return;
+    }
 
     results = await Deck.find({ deckId: deckId }, { __v: 0, _id: 0 }).exec().catch(err => {
         logger.error("models/joints/deck_card_remove/removeCardFromDeck : Erreur de lecture sur la base");
@@ -47,6 +67,17 @@ async function removeCardFromDeck(deckId, cardId, deckPart, userId, callback) {
 
 async function removeAllCardsOfDeck(deckId, callback) {
     logger.debug("Méthode models/joints/deck_card_remove/removeAllCardsOfDeck");
+
+    err = []
+
+    if (typeof deckId === "undefined" || deckId == null || deckId === "") {
+        err.push("Pas de deck fourni.");
+    }
+
+    if (err.length > 0) {
+        callback(err.join('<br/>'));
+        return;
+    }
 
     this.deleteMany({
         deckId: deckId
